+++ python/rateslib/fx/fx_forwards.py
@@ -4,17 +4,19 @@ import json
 import warnings
 from datetime import datetime, timedelta
 from itertools import product
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 from pandas import DataFrame, Series
 
-from rateslib.calendars import CalInput, add_tenor
+from rateslib.calendars import add_tenor
 from rateslib.curves import Curve, LineCurve, MultiCsaCurve, ProxyCurve
 from rateslib.default import NoInput, PlotOutput, _validate_states, _WithState, plot
 from rateslib.dual import Dual, gradient
-from rateslib.dual.utils import DualTypes, Number
 from rateslib.fx.fx_rates import FXRates
+
+if TYPE_CHECKING:
+    from rateslib.typing import DualTypes, Number, CalInput
 
 """
 .. ipython:: python
