+++ python/rateslib/instruments/base.py
@@ -2,15 +2,17 @@ from __future__ import annotations
 
 import abc
 import warnings
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from pandas import DataFrame, concat, isna
 
 from rateslib.curves import Curves
 from rateslib.default import NoInput
-from rateslib.dual.utils import DualTypes
-from rateslib.instruments.utils import FX, NPV, _get_curves_fx_and_base_maybe_from_solver
+from rateslib.instruments.utils import _get_curves_fx_and_base_maybe_from_solver
 from rateslib.solver import Solver
+
+if TYPE_CHECKING:
+    from rateslib.typing import FX, NPV, DualTypes
 
 
 class BaseMixin:
