+++ python/rateslib/instruments/bonds/securities.py
@@ -3,16 +3,16 @@ from __future__ import annotations
 import abc
 import warnings
 from datetime import datetime, timedelta
+from typing import TYPE_CHECKING
 
 import numpy as np
 from pandas import DataFrame, Series
 
 from rateslib import defaults
-from rateslib.calendars import CalInput, add_tenor, dcf
+from rateslib.calendars import add_tenor, dcf
 from rateslib.curves import Curve, IndexCurve, LineCurve, average_rate, index_left
 from rateslib.default import NoInput, _drb
 from rateslib.dual import Dual, Dual2, gradient, quadratic_eqn
-from rateslib.dual.utils import DualTypes, Number
 from rateslib.fx import FXForwards, FXRates
 from rateslib.instruments.base import BaseMixin
 from rateslib.instruments.bonds.conventions import (
@@ -42,6 +42,9 @@ from rateslib.periods import (
     _maybe_local,
 )
 from rateslib.solver import Solver
+
+if TYPE_CHECKING:
+    from rateslib.typing import DualTypes, Number, CalInput
 
 
 class BondMixin:
