+++ python/rateslib/instruments/rates_multi_ccy.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import warnings
 from datetime import datetime
+from typing import TYPE_CHECKING
 
 from pandas import DataFrame, DatetimeIndex, MultiIndex, Series
 
@@ -9,7 +10,6 @@ from rateslib import defaults
 from rateslib.curves import Curve
 from rateslib.default import NoInput
 from rateslib.dual import Dual, Dual2
-from rateslib.dual.utils import DualTypes
 from rateslib.fx import FXForwards, FXRates, forward_fx
 from rateslib.instruments.base import BaseMixin
 from rateslib.instruments.rates_derivatives import BaseDerivative
@@ -35,6 +35,9 @@ from rateslib.solver import Solver
 # Commercial use of this code, and/or copying and redistribution is prohibited.
 # This code cannot be installed or executed on a corporate computer without a paid licence extension
 # Contact info at rateslib.com if this code is observed outside its intended sphere of use.
+
+if TYPE_CHECKING:
+    from rateslib.typing import DualTypes
 
 
 class FXExchange(Sensitivities, BaseMixin):
