+++ python/rateslib/rs.pyi
@@ -1,13 +1,13 @@
 from collections.abc import Sequence
 from datetime import datetime
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from typing_extensions import Self
 
 from rateslib.calendars import CalTypes
-from rateslib.curves.rs import CurveInterpolator
-from rateslib.dual.utils import DualTypes, Number
-from rateslib.dual.variable import Arr1dF64, Arr2dF64
+
+if TYPE_CHECKING:
+    from rateslib.typing import Arr1dF64, Arr2dF64, CurveInterpolator, DualTypes, Number
 
 class ADOrder:
     Zero: ADOrder
